Docs and Tests
- Moved the tests of `src/msg/model.rs` to `tests/msg/model.rs` to make use of the
purpose of the `tests` directory :D

- Fixed some tests to make all pass
- Added some more docs
